Move dashboard to in-repo-addon
