+++ lib/telling-stories-engine/addon/components/feature-module.js
