+++ lib/telling-stories-engine/addon/routes/stories.js
