+++ lib/telling-stories-engine/addon/templates/application.hbs
@@ -1,0 +1,1 @@
+ember-engine application template<br>{{outlet}}
