+++ lib/telling-stories-engine/addon/templates/components/feature-module.hbs
