+++ lib/telling-stories-engine/addon/templates/stories.hbs
