+++ lib/telling-stories-engine/addon/templates/stories/story.hbs
