+++ lib/telling-stories-engine/index.js
@@ -1,0 +1,21 @@
+/*jshint node:true*/
+
+var EngineAddon = require('ember-engines/lib/engine-addon');
+
+module.exports = EngineAddon.extend({
+  name: 'telling-stories-engine',
+  lazyLoading: false,
+
+  included: function(app) {
+    // see: https://github.com/ember-cli/ember-cli/issues/3718
+    if (typeof app.import !== 'function' && app.app) {
+      app = app.app;
+    }
+
+    this.app = app;
+  },
+
+  isDevelopingAddon() {
+    return true;
+  }
+});
