+++ lib/telling-stories-parser/index.js
@@ -1,0 +1,8 @@
+/*jshint node:true*/
+module.exports = {
+  name: 'telling-stories-parser',
+
+  isDevelopingAddon: function() {
+    return true;
+  }
+};
