+++ lib/telling-stories-parser/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "telling-stories-parser",
+  "keywords": [
+    "ember-addon"
+  ]
+}
